Checkstyle fixed for new stock implementation
