+++ src/main/data_access/stock_market/StockInfoDataOutputObject.java
@@ -1,4 +1,4 @@
-package main.data_access.StockMarket;
+package main.data_access.stock_market;
 
 public record StockInfoDataOutputObject(String ticker, double currentPrice, double meanDailyReturnPct,
                                         double standardDeviationPct) {
