+++ src/main/entity/players/inheritor.java
@@ -1,11 +1,10 @@
 package main.entity.players;
 
-import main.entity.Stocks.Stock;
-import main.entity.players.Player;
+import main.entity.stocks.Stock;
 
 import java.awt.*;
 
-import static main.Constants.Constants.INHERITOR_INIT_MONEY;
+import static main.constants.Constants.INHERITOR_INIT_MONEY;
 
 /**
  * A special type of Player called "Inheritor".
