--- src/main/interface_adapter/StockMarket/StockMarketDataController.java
@@ -1,5 +1,0 @@
-package main.interface_adapter.StockMarket;
-
-// TODO: Implement this when cleaning code
-public class StockMarketDataController {
-}
