--- src/main/interface_adapter/StockMarket/StockPresenter.java
@@ -1,10 +1,0 @@
-package main.interface_adapter.StockMarket;
-
-import main.entity.Stocks.Stock;
-import main.entity.players.Player;
-
-public class StockPresenter {
-    public StockViewModel execute(Stock stock, Player player, boolean allowBuy) {
-        return new StockViewModel(stock, player, allowBuy);
-    }
-}
