--- src/main/interface_adapter/StockMarket/StockSellController.java
@@ -1,19 +1,0 @@
-package main.interface_adapter.StockMarket;
-
-import main.entity.Stocks.Stock;
-import main.entity.players.Player;
-import main.use_case.Stocks.SellStock;
-
-public class StockSellController {
-    public void execute(Player player, Stock stock, int quantity) {
-        if (quantity <= 0) {
-            throw new IllegalArgumentException("Quantity must be greater than 0");
-        }
-        if (player.getStockQuantity(stock) < quantity) {
-            throw new IllegalArgumentException("Insufficient stocks to sell");
-        }
-
-        SellStock sellStock = new SellStock();
-        sellStock.execute(player, stock, quantity);
-    }
-}
