+++ src/main/use_case/EndScreen/EndGame.java
@@ -2,11 +2,10 @@ package main.use_case.EndScreen;
 
 import main.entity.players.Player;
 import main.entity.tiles.PropertyTile;
-import main.entity.Stocks.Stock;
+import main.entity.stocks.Stock;
 
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
