+++ src/test/APITest.java
@@ -1,6 +1,6 @@
 package test;
 
-import main.Constants.Config;
+import main.constants.Config;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
